requirements.txt
cinemagoer==2022.12.27
python-dotenv==0.21.1
flask-restful==0.3.9
flask-cors==3.0.10
jellyfish==0.9.0
pyrogram==2.0.97
aiotinydb==2.0.0
tgcrypto==1.2.5
tinydb==4.7.1